Our work from April 9th to April 10th

+++ RayTracing/Light.java
@@ -3,14 +3,14 @@ package RayTracing;
 public class Light {
 	final Vector position;
 	final RGB rgb;
-	final int spec;
-	final int shadow;
-	final int width;
+	final double spec;
+	final double shadow;
+	final double width;
 
 	Light(
-		int px, int py, int pz,
+		double px, double py, double pz,
 		byte r, byte g, byte b,
-		int spec, int shadow, int width) {
+		double spec, double shadow, double width) {
 
 		this.position = new Vector(px, py, pz);
 		this.rgb = new RGB(r, g, b);
